+++ b2sdk/account_info/abstract.py
@@ -7,7 +7,6 @@
 # License https://www.backblaze.com/using_b2_code.html
 #
 ######################################################################
-
 from abc import abstractmethod
 
 from b2sdk.account_info import exception
@@ -116,7 +115,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     @abstractmethod
     def get_account_id(self):
         """
-        Return account ID or raises MissingAccountData exception.
+        Return account ID or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
 
         :rtype: str
         """
@@ -132,7 +131,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     @abstractmethod
     def get_account_auth_token(self):
         """
-        Return account_auth_token or raises MissingAccountData exception.
+        Return account_auth_token or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
 
         :rtype: str
         """
@@ -140,7 +139,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     @abstractmethod
     def get_api_url(self):
         """
-        Return api_url or raises MissingAccountData exception.
+        Return api_url or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
 
         :rtype: str
         """
@@ -148,7 +147,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     @abstractmethod
     def get_application_key(self):
         """
-        Return application_key or raises MissingAccountData exception.
+        Return application_key or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
 
         :rtype: str
         """
@@ -156,7 +155,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     @abstractmethod
     def get_download_url(self):
         """
-        Return download_url or raises MissingAccountData exception.
+        Return download_url or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
 
         :rtype: str
         """
@@ -164,7 +163,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     @abstractmethod
     def get_realm(self):
         """
-        Return realm or raises MissingAccountData exception.
+        Return realm or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
 
         :rtype: str
         """
@@ -188,7 +187,17 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
         :rtype: dict
         """
 
-    @limit_trace_arguments(only=['self', 'api_url', 'download_url', 'minimum_part_size', 'realm'])
+    @abstractmethod
+    def get_s3_api_url(self):
+        """
+        Return s3_api_url or raises :class:`~b2sdk.v1.exception.MissingAccountData` exception.
+
+        :rtype: str
+        """
+
+    @limit_trace_arguments(
+        only=['self', 'api_url', 'download_url', 'minimum_part_size', 'realm', 's3_api_url']
+    )
     def set_auth_data(
         self,
         account_id,
@@ -198,8 +207,9 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
         minimum_part_size,
         application_key,
         realm,
+        s3_api_url,
         allowed=None,
-        application_key_id=None,
+        application_key_id=None
     ):
         """
         Check permission correctness and stores the results of ``b2_authorize_account``.
@@ -218,6 +228,7 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
         :param str realm: a realm to authorize account in
         :param dict allowed: the structure to use for old account info that was saved without 'allowed'
         :param str application_key_id: application key ID
+        :param str s3_api_url: S3-compatible API URL
 
         .. versionchanged:: 0.1.5
            `account_id_or_app_key_id` renamed to `application_key_id`
@@ -225,16 +236,10 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
         if allowed is None:
             allowed = self.DEFAULT_ALLOWED
         assert self.allowed_is_valid(allowed)
+
         self._set_auth_data(
-            account_id,
-            auth_token,
-            api_url,
-            download_url,
-            minimum_part_size,
-            application_key,
-            realm,
-            allowed,
-            application_key_id,
+            account_id, auth_token, api_url, download_url, minimum_part_size, application_key,
+            realm, s3_api_url, allowed, application_key_id
         )
 
     @classmethod
@@ -258,16 +263,8 @@ class AbstractAccountInfo(metaclass=B2TraceMetaAbstract):
     # TODO: make a decorator for set_auth_data()
     @abstractmethod
     def _set_auth_data(
-        self,
-        account_id,
-        auth_token,
-        api_url,
-        download_url,
-        minimum_part_size,
-        application_key,
-        realm,
-        allowed,
-        application_key_id,
+        self, account_id, auth_token, api_url, download_url, minimum_part_size, application_key,
+        realm, s3_api_url, allowed, application_key_id
     ):
         """
         Actually store the auth data.  Can assume that 'allowed' is present and valid.