+++ b2sdk/account_info/in_memory.py
@@ -55,18 +55,11 @@ class InMemoryAccountInfo(UrlPoolAccountInfo):
         self._download_url = None
         self._minimum_part_size = None
         self._realm = None
+        self._s3_api_url = None
 
     def _set_auth_data(
-        self,
-        account_id,
-        auth_token,
-        api_url,
-        download_url,
-        minimum_part_size,
-        application_key,
-        realm,
-        allowed,
-        application_key_id,
+        self, account_id, auth_token, api_url, download_url, minimum_part_size, application_key,
+        realm, s3_api_url, allowed, application_key_id
     ):
         self._account_id = account_id
         self._application_key_id = application_key_id
@@ -76,6 +69,7 @@ class InMemoryAccountInfo(UrlPoolAccountInfo):
         self._minimum_part_size = minimum_part_size
         self._application_key = application_key
         self._realm = realm
+        self._s3_api_url = s3_api_url
         self._allowed = allowed
 
     def refresh_entire_bucket_name_cache(self, name_id_iterable):
@@ -126,3 +120,7 @@ class InMemoryAccountInfo(UrlPoolAccountInfo):
     @_raise_missing_if_result_is_none
     def get_allowed(self):
         return self._allowed
+
+    @_raise_missing_if_result_is_none
+    def get_s3_api_url(self):
+        return self._s3_api_url