+++ b2sdk/v0/__init__.py
@@ -8,7 +8,7 @@
 #
 ######################################################################
 
-from b2sdk.v1 import *
+from b2sdk.v1 import *  # noqa
 from b2sdk.v0.account_info import AbstractAccountInfo, InMemoryAccountInfo, UrlPoolAccountInfo, SqliteAccountInfo
 from b2sdk.v0.api import B2Api
 from b2sdk.v0.bucket import Bucket