+++ test/unit/account_info/__init__.py
@@ -1,11 +1,9 @@
 ######################################################################
 #
-# File: test/unit/v0/apiver/apiver_deps.py
+# File: test/unit/account_info/__init__.py
 #
-# Copyright 2020 Backblaze Inc. All Rights Reserved.
+# Copyright 2021 Backblaze Inc. All Rights Reserved.
 #
 # License https://www.backblaze.com/using_b2_code.html
 #
 ######################################################################
-
-from b2sdk.v0 import *  # noqa