+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Calculate S3-compatible API URL base on API URL
+
 ### Fixed
 * Exclude packages inside the test package when installing
 
@@ -13,7 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Add `dependabot.yml`
-* Add support for SSE-B2 server-side encryption mode
 
 ### Changed
 * Add upper version limit for the requirements